Added .gitignore.
The .gitignore file ignores by-products of running LaTeX, both by itself
and through Emacs (hence Kokkos_PG.synctex.gz).

NOTE: Kokkos_PG.pdf is in .gitignore.  This is the name of the PDF file
which is the Kokkos Programming Guide.  This means that by default, you
CANNOT check the finished PDF into the repository.  If for some reason
you need to check in the finished PDF, you must first change .gitignore
and remove the Kokkos_PG.pdf line.