Replaced [=] with KOKKOS_LAMBDA; explained Kokkos::single better.
I added an explanation of the KOKKOS_LAMBDA macro.  If CUDA is OFF, it
just turns into [=].  If CUDA is ON, it may do something special to make
the lambda work with CUDA.

I also improved the explanation of Kokkos::single.  In particular,
Kokkos::single's lambda is NOT allowed to modify any captured variable
by reference.